Assets 72092 (#166)
* move calendar functions to main details block

* fix small issues

- add null check to release tier property
- refactor logic when checking for bad view starts to account for non-US
  time zones

* remove commented code
